Extract portfolio view component per naming scheme (migration step 3).
Move port route UI into portfolio-view.js so page.js stays a thin Next.js adapter.
